get rif of skip_not_selenium or skip_not_splinter
