+++ tests/test_page.py
@@ -7,10 +7,6 @@ import random
 import pytest
 
 from pypom import Page
-from .util import (
-    skip_not_selenium,
-    skip_not_splinter,
-)
 
 
 def test_base_url(base_url, page):
@@ -98,145 +94,6 @@ def test_wait_for_page_timeout(base_url, driver):
 
 def test_wait_for_page_empty_base_url(driver):
     assert isinstance(Page(driver).wait_for_page_to_load(), Page)
-
-
-def test_find_element_selenium(page, driver, driver_interface):
-    skip_not_selenium(driver_interface)
-
-    locator = (str(random.random()), str(random.random()))
-    page.find_element(*locator)
-    driver.find_element.assert_called_once_with(*locator)
-
-
-def test_find_element_splinter(page, driver, driver_interface, splinter_strategy):
-    skip_not_splinter(driver_interface)
-
-    locator = (splinter_strategy, str(random.random()))
-    page.find_element(*locator)
-    getattr(page.driver, 'find_by_{0}'.format(splinter_strategy)).assert_called_once_with(locator[1])
-
-
-def test_find_elements_selenium(page, driver, driver_interface):
-    skip_not_selenium(driver_interface)
-
-    locator = (str(random.random()), str(random.random()))
-    page.find_elements(*locator)
-    driver.find_elements.assert_called_once_with(*locator)
-
-
-def test_find_elements_splinter(page, driver, driver_interface, splinter_strategy):
-    skip_not_splinter(driver_interface)
-
-    locator = (splinter_strategy, str(random.random()))
-    page.find_elements(*locator)
-    getattr(page.driver, 'find_by_{0}'.format(splinter_strategy)).assert_called_once_with(locator[1])
-
-
-def test_is_element_present_selenium(page, driver, driver_interface):
-    skip_not_selenium(driver_interface)
-
-    locator = (str(random.random()), str(random.random()))
-    assert page.is_element_present(*locator)
-    driver.find_element.assert_called_once_with(*locator)
-
-
-def test_is_element_present_splinter(page, driver, driver_interface, splinter_strategy):
-    skip_not_splinter(driver_interface)
-
-    locator = (splinter_strategy, str(random.random()))
-    from splinter.element_list import ElementList
-    from mock import Mock
-    page.driver.configure_mock(**{'find_by_{0}.return_value'.format(splinter_strategy): ElementList([Mock()])})
-    assert page.is_element_present(*locator)
-    getattr(page.driver, 'find_by_{0}'.format(splinter_strategy)).assert_called_once_with(locator[1])
-
-
-def test_is_element_present_not_present_selenium(page, driver, driver_interface):
-    skip_not_selenium(driver_interface)
-
-    locator = (str(random.random()), str(random.random()))
-    from selenium.common.exceptions import NoSuchElementException
-    driver.find_element.side_effect = NoSuchElementException()
-    assert not page.is_element_present(*locator)
-    driver.find_element.assert_called_once_with(*locator)
-
-
-def test_is_element_present_not_present_splinter(page, driver, driver_interface, splinter_strategy):
-    skip_not_splinter(driver_interface)
-
-    locator = (splinter_strategy, str(random.random()))
-    from splinter.element_list import ElementList
-    page.driver.configure_mock(**{'find_by_{0}.return_value'.format(splinter_strategy): ElementList([])})
-    assert not page.is_element_present(*locator)
-    getattr(page.driver, 'find_by_{0}'.format(splinter_strategy)).assert_called_once_with(locator[1])
-
-
-def test_is_element_displayed_selenium(page, driver, driver_interface):
-    skip_not_selenium(driver_interface)
-
-    locator = (str(random.random()), str(random.random()))
-    assert page.is_element_displayed(*locator)
-    driver.find_element.assert_called_once_with(*locator)
-
-
-def test_is_element_displayed_splinter(page, driver, driver_interface, splinter_strategy):
-    skip_not_splinter(driver_interface)
-
-    locator = (splinter_strategy, str(random.random()))
-
-    from mock import PropertyMock
-    visible_mock = PropertyMock(return_value=True)
-    page.driver.configure_mock(**{'find_by_{0}.return_value.first.visible'.format(splinter_strategy): visible_mock})
-    type(getattr(page.driver, 'find_by_{0}'.format(splinter_strategy)).return_value.first).visible = visible_mock
-    assert page.is_element_displayed(*locator)
-    getattr(page.driver, 'find_by_{0}'.format(splinter_strategy)).assert_called_once_with(locator[1])
-    visible_mock.assert_called_with()
-
-
-def test_is_element_displayed_not_present_selenium(page, driver, driver_interface):
-    skip_not_selenium(driver_interface)
-
-    locator = (str(random.random()), str(random.random()))
-    from selenium.common.exceptions import NoSuchElementException
-    driver.find_element.side_effect = NoSuchElementException()
-    assert not page.is_element_displayed(*locator)
-    driver.find_element.assert_called_once_with(*locator)
-    driver.find_element.is_displayed.assert_not_called()
-
-
-def test_is_element_displayed_not_present_splinter(page, driver, driver_interface, splinter_strategy):
-    skip_not_splinter(driver_interface)
-
-    locator = (splinter_strategy, str(random.random()))
-    from splinter.element_list import ElementList
-    page.driver.configure_mock(**{'find_by_{0}.return_value'.format(splinter_strategy): ElementList([])})
-    assert not page.is_element_displayed(*locator)
-    getattr(page.driver, 'find_by_{0}'.format(splinter_strategy)).assert_called_once_with(locator[1])
-
-
-def test_is_element_displayed_not_displayed_selenium(page, driver, driver_interface):
-    skip_not_selenium(driver_interface)
-
-    locator = (str(random.random()), str(random.random()))
-    element = driver.find_element()
-    element.is_displayed.return_value = False
-    assert not page.is_element_displayed(*locator)
-    driver.find_element.assert_called_with(*locator)
-    element.is_displayed.assert_called_once_with()
-
-
-def test_is_element_displayed_not_displayed_splinter(page, driver, driver_interface, splinter_strategy):
-    skip_not_splinter(driver_interface)
-
-    locator = (splinter_strategy, str(random.random()))
-
-    from mock import PropertyMock
-    visible_mock = PropertyMock(return_value=False)
-    page.driver.configure_mock(**{'find_by_{0}.return_value.first.visible'.format(splinter_strategy): visible_mock})
-    type(getattr(page.driver, 'find_by_{0}'.format(splinter_strategy)).return_value.first).visible = visible_mock
-    assert not page.is_element_displayed(*locator)
-    getattr(page.driver, 'find_by_{0}'.format(splinter_strategy)).assert_called_once_with(locator[1])
-    visible_mock.assert_called_with()
 
 
 def test_bwc_selenium(page, driver_interface):
